+++ pyoffers/models/__init__.py
@@ -1,5 +1,6 @@
 # coding: utf-8
 from .advertiser import Advertiser, AdvertiserManager  # noqa
+from .affiliate import Affiliate, AffiliateManager, AffiliateUser, AffiliateUserManager  # noqa
 from .conversion import Conversion, ConversionManager  # noqa
 from .core import ModelManager  # noqa
 from .country import Country, CountryManager  # noqa
@@ -9,5 +10,12 @@ from .raw_log import RawLogManager  # noqa
 
 
 MODEL_MANAGERS = (
-    AdvertiserManager, ConversionManager, CountryManager, GoalManager, OfferManager, RawLogManager
+    AdvertiserManager,
+    ConversionManager,
+    CountryManager,
+    GoalManager,
+    OfferManager,
+    RawLogManager,
+    AffiliateManager,
+    AffiliateUserManager
 )